+++ CHANGELOG.md
@@ -5,6 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.2
 ### Fixed
 - fix watch error caused by missing folder
+- added `surge` and `github` environment settings
+- updated `npm run deploy` to use the `github` environment
+
+## 0.1.2
+### Changed
+- updated form encoding to use `encodeURIComponent`
 
 ## 0.1.1
 ### Added